requirements.txt
python-telegram-bot==13.7
Pillow==8.2.0